+++ examples/nuxt-openai/pages/use-chat-tools/index.vue
@@ -13,65 +13,116 @@ const { input, handleSubmit, messages, addToolResult } = useChat({
     }
   },
 });
+
+const messageList = computed(() => messages.value); // computer property for type inference
 </script>
 
 <template>
   <div class="flex flex-col w-full max-w-md py-24 mx-auto stretch">
-    <div v-for="m in messages" :key="m.id" class="whitespace-pre-wrap">
-      <strong>{{ `${m.role}: ` }}</strong>
-      {{ m.content }}
-      <template
-        v-for="toolInvocation in m.toolInvocations"
-        :key="toolInvocation.toolCallId"
-      >
-        <!-- example of pre-rendering streaming tool calls -->
-        <pre v-if="toolInvocation.state === 'partial-call'">
-          {{ JSON.stringify(toolInvocation, null, 2) }}
-        </pre>
-
-        <!-- render confirmation tool (client-side tool with user interaction) -->
-        <div
-          v-else-if="toolInvocation.toolName === 'askForConfirmation'"
-          class="text-gray-500"
-        >
-          {{ toolInvocation.args.message }}
-          <div class="flex gap-2">
-            <b v-if="'result' in toolInvocation">{{ toolInvocation.result }}</b>
-            <template v-else>
-              <button
-                class="px-4 py-2 font-bold text-white bg-blue-500 rounded hover:bg-blue-700"
-                @click="
-                  addToolResult({
-                    toolCallId: toolInvocation.toolCallId,
-                    result: 'Yes, confirmed.',
-                  })
-                "
+    <div
+      v-for="message in messageList"
+      :key="message.id"
+      class="whitespace-pre-wrap"
+    >
+      <strong>{{ `${message.role}: ` }}</strong>
+      <template v-for="part in message.parts">
+        <template v-if="part.type === 'text'">
+          {{ part.text }}
+        </template>
+        <template v-else-if="part.type === 'tool-invocation'">
+          <template
+            v-if="part.toolInvocation.toolName === 'askForConfirmation'"
+          >
+            <template v-if="part.toolInvocation.state === 'call'">
+              <div
+                :key="part.toolInvocation.toolCallId"
+                className="text-gray-500"
               >
-                Yes
-              </button>
-              <button
-                class="px-4 py-2 font-bold text-white bg-red-500 rounded hover:bg-red-700"
-                @click="
-                  addToolResult({
-                    toolCallId: toolInvocation.toolCallId,
-                    result: 'No, denied',
-                  })
-                "
-              >
-                No
-              </button>
+                {{ part.toolInvocation.args.message }}
+                <div className="flex gap-2">
+                  <button
+                    class="px-4 py-2 font-bold text-white bg-blue-500 rounded hover:bg-blue-700"
+                    @click="
+                      addToolResult({
+                        toolCallId: part.toolInvocation.toolCallId,
+                        result: 'Yes, confirmed.',
+                      })
+                    "
+                  >
+                    Yes
+                  </button>
+                  <button
+                    class="px-4 py-2 font-bold text-white bg-red-500 rounded hover:bg-red-700"
+                    @click="
+                      addToolResult({
+                        toolCallId: part.toolInvocation.toolCallId,
+                        result: 'No, denied',
+                      })
+                    "
+                  >
+                    No
+                  </button>
+                </div>
+              </div>
             </template>
-          </div>
-        </div>
-
-        <!-- other tools -->
-        <div v-else class="text-gray-500">
-          <template v-if="'result' in toolInvocation">
-            Tool call {{ `${toolInvocation.toolName}: ` }}
-            {{ toolInvocation.result }}
+            <template v-if="part.toolInvocation.state === 'result'">
+              <div
+                :key="part.toolInvocation.toolCallId"
+                className="text-gray-500"
+              >
+                Location access allowed: {{ part.toolInvocation.result }}
+              </div>
+            </template>
           </template>
-          <template v-else> Calling {{ toolInvocation.toolName }}... </template>
-        </div>
+
+          <template v-if="part.toolInvocation.toolName === 'getLocation'">
+            <template v-if="part.toolInvocation.state === 'call'">
+              <div
+                :key="part.toolInvocation.toolCallId"
+                className="text-gray-500"
+              >
+                Getting location...
+              </div>
+            </template>
+            <template v-if="part.toolInvocation.state === 'result'">
+              <div
+                :key="part.toolInvocation.toolCallId"
+                className="text-gray-500"
+              >
+                Location: {{ part.toolInvocation.result }}
+              </div>
+            </template>
+          </template>
+
+          <template
+            v-if="part.toolInvocation.toolName === 'getWeatherInformation'"
+          >
+            <template v-if="part.toolInvocation.state === 'partial-call'">
+              <pre :key="part.toolInvocation.toolCallId">
+                {{ JSON.stringify(part.toolInvocation, null, 2) }}
+              </pre>
+            </template>
+            <template v-if="part.toolInvocation.state === 'call'">
+              <div
+                :key="part.toolInvocation.toolCallId"
+                className="text-gray-500"
+              >
+                Getting weather information for
+                {{ part.toolInvocation.args.city }}...
+              </div>
+            </template>
+            <template v-if="part.toolInvocation.state === 'result'">
+              <div
+                :key="part.toolInvocation.toolCallId"
+                className="text-gray-500"
+              >
+                Weather in {{ part.toolInvocation.args.city }}:
+                {{ part.toolInvocation.result }}
+              </div>
+            </template>
+          </template>
+        </template>
+        <br />
       </template>
     </div>