+++ packages/ai/core/generate-text/generate-text.ts
@@ -198,7 +198,7 @@ A function that attempts to repair a tool call that failed to parse.
     onStepFinish?: (event: StepResult<TOOLS>) => Promise<void> | void;
 
     /**
-     * @internal For test use only. May change without notice.
+     * Internal. For test use only. May change without notice.
      */
     _internal?: {
       generateId?: IDGenerator;
@@ -292,7 +292,7 @@ A function that attempts to repair a tool call that failed to parse.
             messages: stepInputMessages,
           },
           modelSupportsImageUrls: model.supportsImageUrls,
-          modelSupportsUrl: model.supportsUrl,
+          modelSupportsUrl: model.supportsUrl?.bind(model), // support 'this' context
         });
 
         currentModelResponse = await retry(() =>