+++ packages/ai/core/generate-text/stream-text.ts
@@ -288,7 +288,7 @@ Callback that is called when each step (LLM call) is finished, including interme
     onStepFinish?: (event: StepResult<TOOLS>) => Promise<void> | void;
 
     /**
-@internal For test use only. May change without notice.
+Internal. For test use only. May change without notice.
      */
     _internal?: {
       now?: () => number;
@@ -865,7 +865,7 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
               messages: stepInputMessages,
             },
             modelSupportsImageUrls: model.supportsImageUrls,
-            modelSupportsUrl: model.supportsUrl,
+            modelSupportsUrl: model.supportsUrl?.bind(model), // support 'this' context
           });
 
           const mode = {