+++ packages/ai/core/registry/custom-provider.ts
@@ -1,4 +1,9 @@
-import { EmbeddingModelV1, LanguageModelV1 } from '@ai-sdk/provider';
+import {
+  EmbeddingModelV1,
+  LanguageModelV1,
+  ImageModelV1,
+  ProviderV1,
+} from '@ai-sdk/provider';
 import { Provider } from '../types';
 import { NoSuchModelError } from '@ai-sdk/provider';
 
@@ -13,17 +18,29 @@ import { NoSuchModelError } from '@ai-sdk/provider';
  *
  * @throws {NoSuchModelError} Throws when a requested model is not found and no fallback provider is available.
  */
-export function experimental_customProvider({
+export function customProvider<
+  LANGUAGE_MODELS extends Record<string, LanguageModelV1>,
+  EMBEDDING_MODELS extends Record<string, EmbeddingModelV1<string>>,
+  IMAGE_MODELS extends Record<string, ImageModelV1>,
+>({
   languageModels,
   textEmbeddingModels,
+  imageModels,
   fallbackProvider,
 }: {
-  languageModels?: Record<string, LanguageModelV1>;
-  textEmbeddingModels?: Record<string, EmbeddingModelV1<string>>;
-  fallbackProvider?: Provider;
-}): Provider {
+  languageModels?: LANGUAGE_MODELS;
+  textEmbeddingModels?: EMBEDDING_MODELS;
+  imageModels?: IMAGE_MODELS;
+  fallbackProvider?: ProviderV1;
+}): Provider & {
+  languageModel(modelId: ExtractModelId<LANGUAGE_MODELS>): LanguageModelV1;
+  textEmbeddingModel(
+    modelId: ExtractModelId<EMBEDDING_MODELS>,
+  ): EmbeddingModelV1<string>;
+  imageModel(modelId: ExtractModelId<IMAGE_MODELS>): ImageModelV1;
+} {
   return {
-    languageModel(modelId: string): LanguageModelV1 {
+    languageModel(modelId: ExtractModelId<LANGUAGE_MODELS>): LanguageModelV1 {
       if (languageModels != null && modelId in languageModels) {
         return languageModels[modelId];
       }
@@ -35,7 +52,9 @@ export function experimental_customProvider({
       throw new NoSuchModelError({ modelId, modelType: 'languageModel' });
     },
 
-    textEmbeddingModel(modelId: string): EmbeddingModelV1<string> {
+    textEmbeddingModel(
+      modelId: ExtractModelId<EMBEDDING_MODELS>,
+    ): EmbeddingModelV1<string> {
       if (textEmbeddingModels != null && modelId in textEmbeddingModels) {
         return textEmbeddingModels[modelId];
       }
@@ -46,5 +65,27 @@ export function experimental_customProvider({
 
       throw new NoSuchModelError({ modelId, modelType: 'textEmbeddingModel' });
     },
+
+    imageModel(modelId: ExtractModelId<IMAGE_MODELS>): ImageModelV1 {
+      if (imageModels != null && modelId in imageModels) {
+        return imageModels[modelId];
+      }
+
+      if (fallbackProvider?.imageModel) {
+        return fallbackProvider.imageModel(modelId);
+      }
+
+      throw new NoSuchModelError({ modelId, modelType: 'imageModel' });
+    },
   };
 }
+
+/**
+ * @deprecated Use `customProvider` instead.
+ */
+export const experimental_customProvider = customProvider;
+
+type ExtractModelId<MODELS extends Record<string, unknown>> = Extract<
+  keyof MODELS,
+  string
+>;