+++ packages/solid/src/use-chat.ui.test.tsx
@@ -103,14 +103,8 @@ describe('prepareRequestBody', () => {
 
 describe('file attachments with data url', () => {
   const TestComponent = () => {
-    const {
-      messages,
-      handleSubmit,
-      handleInputChange,
-      isLoading,
-      input,
-      setInput,
-    } = useChat();
+    const { messages, handleSubmit, handleInputChange, isLoading, input } =
+      useChat();
 
     const [attachments, setAttachments] = createSignal<FileList | undefined>();
     let fileInputRef: HTMLInputElement | undefined;
@@ -228,6 +222,7 @@ describe('file attachments with data url', () => {
                   url: 'data:text/plain;base64,dGVzdCBmaWxlIGNvbnRlbnQ=',
                 },
               ],
+              parts: [{ text: 'Message with text attachment', type: 'text' }],
             },
           ],
         });
@@ -466,6 +461,7 @@ describe('data protocol stream', () => {
               createdAt: expect.any(Date),
               role: 'assistant',
               content: 'Hello, world.',
+              parts: [{ text: 'Hello, world.', type: 'text' }],
             },
             options: {
               finishReason: 'stop',
@@ -604,6 +600,7 @@ describe('text stream', () => {
               createdAt: expect.any(Date),
               role: 'assistant',
               content: 'Hello, world.',
+              parts: [{ text: 'Hello, world.', type: 'text' }],
             },
             options: {
               finishReason: 'unknown',
@@ -1287,7 +1284,13 @@ describe('reload', () => {
 
         expect(await call(1).getRequestBodyJson()).toStrictEqual({
           id: expect.any(String),
-          messages: [{ content: 'hi', role: 'user' }],
+          messages: [
+            {
+              content: 'hi',
+              role: 'user',
+              parts: [{ text: 'hi', type: 'text' }],
+            },
+          ],
           data: { 'test-data-key': 'test-data-value' },
           'request-body-key': 'request-body-value',
         });