+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,49 @@
 
 ### Patch Changes
 
+- Updated dependencies [bcc61d4]
+  - ai@4.1.20
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.1.19
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [6a1acfe]
+  - ai@4.1.18
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [5af8cdb]
+  - ai@4.1.17
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [7e299a4]
+  - ai@4.1.16
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [d89c3b9]
+- Updated dependencies [d89c3b9]
+- Updated dependencies [08f54fc]
+  - ai@4.1.15
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [ca89615]
   - ai@4.1.14