+++ packages/fireworks/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ai-sdk/fireworks
 
+## 0.1.8
+
+### Patch Changes
+
+- caf7d6f: fix (provider): extends providerv1 spec
+- d89c3b9: feat (provider): add image model support to provider specification
+- Updated dependencies [d89c3b9]
+  - @ai-sdk/provider@1.0.7
+  - @ai-sdk/openai-compatible@0.1.8
+  - @ai-sdk/provider-utils@2.1.6
+
 ## 0.1.7
 
 ### Patch Changes