+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/groq
 
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies [d89c3b9]
+  - @ai-sdk/provider@1.0.7
+  - @ai-sdk/provider-utils@2.1.6
+
 ## 1.1.6
 
 ### Patch Changes